agent/src/workflows/sre_workflow.py
"""
Main SRE Workflow using AutoGen GroupChat.

This is the primary orchestrator that manages multi-agent conversations
for Kubernetes incident response and remediation.
"""

from __future__ import annotations

import asyncio
from typing import Any

import structlog
from autogen import ConversableAgent, GroupChat, GroupChatManager

from ..agents.analysis import AnalysisAgent
from ..config import get_settings

logger = structlog.get_logger()


class SREWorkflow:
    """
    Main SRE workflow orchestrator using AutoGen GroupChat.

    This class manages the multi-agent conversation flow:
    1. Analysis Agent diagnoses the issue
    2. Recommendation Agent suggests actions
    3. Guard Agent validates safety
    4. Approval Agent makes final decision
    5. Execution Agent implements approved actions
    """

    def __init__(self):
        self.settings = get_settings()
        self.agents = self._create_agents()
        self.group_chat = self._create_group_chat()
        self.manager = self._create_manager()

    def _create_agents(self) -> dict[str, ConversableAgent]:
        """Create all agents for the workflow."""

        # Create analysis agent
        analysis_agent = AnalysisAgent(
            name="analysis_agent",
            llm_config={
                "model": "gpt-4",
                "temperature": 0.3,
                "timeout": 60,
            },
        )

        # TODO: Create other agents
        # recommendation_agent = RecommendationAgent(...)
        # guard_agent = GuardAgent(...)
        # approval_agent = ApprovalAgent(...)
        # execution_agent = ExecutionAgent(...)

        # For now, create placeholder agents
        recommendation_agent = ConversableAgent(
            name="recommendation_agent",
            system_message="You suggest remediation actions based on analysis.",
            llm_config={"model": "gpt-4", "temperature": 0.2},
        )

        guard_agent = ConversableAgent(
            name="guard_agent",
            system_message="You validate safety of proposed actions.",
            llm_config={"model": "gpt-3.5-turbo", "temperature": 0.0},
        )

        approval_agent = ConversableAgent(
            name="approval_agent",
            system_message="You make final approval decisions for actions.",
            llm_config={"model": "gpt-4", "temperature": 0.1},
        )

        return {
            "analysis": analysis_agent,
            "recommendation": recommendation_agent,
            "guard": guard_agent,
            "approval": approval_agent,
        }

    def _create_group_chat(self) -> GroupChat:
        """Create the multi-agent group chat."""
        return GroupChat(
            agents=list(self.agents.values()),
            messages=[],
            max_round=10,  # Maximum conversation rounds
            speaker_selection_method="auto",  # Let AutoGen manage turn-taking
        )

    def _create_manager(self) -> GroupChatManager:
        """Create the group chat manager."""
        return GroupChatManager(
            groupchat=self.group_chat,
            llm_config={
                "model": "gpt-4",
                "temperature": 0.0,
                "timeout": 60,
            },
        )

    async def process_incident(
        self, event_data: dict[str, Any], namespace: str, resource_name: str
    ) -> dict[str, Any]:
        """
        Process a Kubernetes incident using multi-agent workflow.

        Args:
            event_data: Kubernetes event data
            namespace: K8s namespace
            resource_name: Resource name

        Returns:
            Decision result with recommended actions
        """
        logger.info(
            "Starting SRE workflow", namespace=namespace, resource=resource_name
        )

        # Initial message to start the conversation
        initial_message = f"""
        Kubernetes Incident Detected:

        Namespace: {namespace}
        Resource: {resource_name}
        Event Data: {event_data}

        Please analyze this incident and recommend appropriate actions.
        Analysis Agent should start by examining the events and symptoms.
        """

        try:
            # Start the multi-agent conversation
            result = await self.manager.a_initiate_chat(
                self.agents["analysis"], message=initial_message, max_turns=10
            )

            # Process the conversation result
            return self._extract_decision(result)

        except Exception as e:
            logger.error("Workflow processing failed", error=str(e))
            return {
                "decision": "error",
                "confidence": 0.0,
                "recommended_actions": [],
                "reasoning": f"Workflow failed: {str(e)}",
            }

    def _extract_decision(self, conversation_result: Any) -> dict[str, Any]:
        """Extract final decision from conversation result."""

        # TODO: Implement proper result extraction
        # This should parse the conversation and extract:
        # - Final decision (approve/reject/human_review)
        # - Confidence level
        # - Recommended actions
        # - Reasoning

        return {
            "decision": "human_review",
            "confidence": 0.5,
            "recommended_actions": [],
            "reasoning": "Workflow processing completed, requires human review",
        }
