agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-agent"
version = "0.1.0"
description = "AutoGen-based SRE Agent for Kubernetes operations"
readme = "README.md"
authors = [
    {name = "K8s Lovers Korea"}
]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "autogen-agentchat>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "kubernetes>=28.1.0",
    "prometheus-client>=0.19.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]
azure = [
    "azure-ai-inference>=1.0.0b1",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
]
all = [
    "sre-agent[dev,azure]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.isort]
known-first-party = ["agents", "tools", "workflows", "guards", "api"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--strict-markers --strict-config"
